requirements.txt
python-magic
requests
pytumblr
nudepy
